3rd pers camera stuff, health damage overlays
Also modified movements speeds, improved leaning (the light saber needs
leaning and 3rd person has no leaning yet). Started working on a better
over the should 3rd camera for playing in third person view.

+++ code/game/g_weapon.cpp
@@ -3135,7 +3135,7 @@ void SubtractLeanOfs(const gentity_t *const ent, vec3_t point)
 			vec3_t	right;
 			//add leaning offset
 			AngleVectors( ent->client->ps.viewangles, NULL, right, NULL );
-			VectorMA( point, ent->client->ps.leanofs*-1, right, point );
+			VectorMA( point, (float)ent->client->ps.leanofs*-1, right, point );
 		}
 	}
 }
